Moved TemplateTagLoading cases into test_custom.
